+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 ##### Enhancements
 
-* None.
+* Allow specifying a `newLineScalarStyle` for encoding string scalars with newlines when using `YAMLEncoder`
+  [Tejas Sharma](https://github.com/tejassharma96)
+  [#405](https://github.com/jpsim/Yams/issues/405)
 
 ##### Bug Fixes